clean: move non page files to misc folder
